Improve PyInstaller compatiblity.
PyInstaller saves the original value to *_ORIG, then modifies the search path so that the bundled libraries
are found first by the bundled code.
But if your code executes a system program, you often do not want that this system program loads your bundled
libraries (that are maybe not compatible with your system program) - it rather should load the correct libraries
from the system locations like it usually does.
Thus you need to restore the original path before creating the subprocess with the system program
https://github.com/pyinstaller/pyinstaller/blob/483c819d6a256b58db6740696a901bd41c313f0c/doc/runtime-information.rst#ld_library_path--libpath-considerations